Move auth package to vicadmin
