+++ cmd/vicadmin/pkg/auth/auth.go
